feat : page extensions 추가
