+++ next.config.mjs
@@ -3,6 +3,7 @@ import {withSentryConfig} from "@sentry/nextjs";
 const nextConfig = {
   reactStrictMode: true,
   transpilePackages: ["rc-util", "@ant-design", "kitchen-flow-editor", "@ant-design/pro-editor", "zustand", "leva", "antd", "rc-pagination", "rc-picker"],
+  pageExtensions: ["page.tsx", "page.ts", "page.jsx", "page.js"],
 };
 
 export default withSentryConfig(nextConfig, {
